requirements_gpu.txt
gymnasium>=0.29.1
stable-baselines3>=2.3.2
stable-retro>=0.9.3
opencv-python-headless
torch>=2.3.0
